add the ability to designate sample labels independently
addGroupCoverages automatically assumes that sample labels come from cellColData$Sample which assumes that each arrow file represents an individual sample. This is not the case for applications where different samples are mixed together into the same GEM reaction as is the case for multiplexing based on sample genotype or lipid barcode.