params must be a list, not a vector
the params section of ArchRProj@embeddings needs to be a list rather than a vector. If a vector, when you try to pass a vector to dimsToUse, you cant have a vector of a vector so this gets converted to dimsToUse1, dimsToUse2, dimsToUse3 etc instead of dimsToUse = c(1,2,3). Discussed in https://github.com/GreenleafLab/ArchR/issues/1379#issuecomment-1099899524