fix divide by zero
as reported here - https://github.com/GreenleafLab/ArchR/pull/1791
There is a bug if the first LSI feature (idx = 1) is part of the top/variable features used for LSI because this would lead to a divide by 0 when trying to obtain the width. Since TileMatrix and PeakMatrix are the only currently supported options for LSI, this update makes it so that if it is a PeakMatrix that was used for LSI, then the end coordinates are used and if it was a TileMatrix then the `rD$tileSize` value is used.