set bcPass as a character vector
I've sporadically received a cryptic error from createArrowFiles() that seems to boil down to an S4Vectors issue with the match() function and the BStringSet bcPass object (shown below). This should be fixed by this commit but its a sporadic error and hard to test.

```
<simpleError in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': error in evaluating the argument 'x' in selecting a method for function 'which': no met
hod found for function 'match' and signature XRawList, XRawList>
```